Rewrote HypnoDisc class to use std::vector for heavy lifting.
The constructor for the HypnoDisc has been reduced to an initializer list. The setLength method has been removed, making live size changes impossible but this is not a likely usecase anyway. Various iterations over the ledStates vector have been implemented using iterators and the code for clockwiseSpin has been simplified to use clockwiseWipe for most of the work.

Latching and writing out ledStates still needs improvement and cleanup.